feat: refactor run-ios to separate files, export more utilities
